Fix 32-bit overflow in fastq_eestats ee_start() for long reads
ee_start() computed pos * ((resolution * (pos + 1)) + 2) / 2 entirely in
32-bit int (every operand was int) and only widened the result to the
int64_t return type. With the default resolution (1000) the value grows
~500*pos^2 and exceeds INT_MAX once pos is around 2074, so for reads
longer than ~2074 bp (routine for PacBio/Nanopore and merged amplicons)
the result overflowed. That overflowed value both sizes ee_length_table
and indexes it, so once it overflowed the allocation size and the indices
disagreed, giving out-of-bounds heap reads and writes.

Take pos as int64_t so the arithmetic is performed in 64-bit. All four
call sites already pass int64_t lengths, which were previously narrowed to
the int parameter, so this also removes that narrowing.

+++ src/eestats.cc
@@ -124,7 +124,7 @@ auto q2p(int const quality_value) -> double
   return std::pow(base, -quality_value / base);
 }
 
-auto ee_start(int const pos, int const resolution) -> int64_t
+auto ee_start(int64_t const pos, int const resolution) -> int64_t
 {
   return pos * ((resolution * (pos + 1)) + 2) / 2;
 }
